Update default return value in ExtractModelTexts method
An adjustment was made to the ExtractModelTexts method in the GeminiTextExtractor program. Specifically, the default return value has been changed from a new List<string> to an empty array when "chunks" is null. The change was made to return a less complex type when no matching chunks are found in the provided JSON object.